requirements.txt
# Robinhood Trading Bot - Python Dependencies

# Core Trading Library
robin-stocks==3.0.5

# Web Framework for API
flask==3.0.0
flask-cors==4.0.0

# Security & Authentication
python-dotenv==1.0.0
cryptography==41.0.7

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Rate Limiting & Safety
ratelimit==2.2.1

# Logging
colorlog==6.8.0

# Environment Management
pyotp==2.9.0  # For 2FA/MFA support
